- jvm `@get:JvmName("getVtxBuffer")` - `TextureID` to `Long`

+++ core/src/main/kotlin/imgui/helpers.kt
@@ -36,7 +36,7 @@ typealias ID = Int
 /** ImTexture: user data for renderer backend to identify a texture [Compile-time configurable type]
  *  - To use something else than an opaque void* pointer: override with e.g. '#define ImTextureID MyTextureType*' in your imconfig.h file.
  *  - This can be whatever to you want it to be! read the FAQ about ImTextureID for details. */
-typealias TextureID = Int
+typealias TextureID = Long
 
 /** Return false = pass
  *
